[age] allow 1Password sessions to work
turns out an explicit account name means chezmoi is looking for a
session variable that doesn't exist--op 2.0 stores the session token in
an env var with the account UUID instead of the account name.

by removing the account name and just using the item UUID, the existing
session gets reused and we no longer have to enter the vault password
every single time we run any chezmoi command.